Clan only posts now working properly

+++ comet/modules/posts/controllers/admin.php
@@ -61,7 +61,7 @@ class Admin extends Backend_Controller {
 			if (!empty($_FILES['userfile']['name']))
 			{
 				$config['upload_path']   = $this->folder_path;
-				$config['allowed_types'] = 'gif|jpg|png';
+				$config['allowed_types'] = 'gif|jpg|png|bmp|jpeg';
 				$config['max_size']      = '0';
 				$config['max_width']     = '0';
 				$config['max_height']    = '0';
@@ -109,6 +109,8 @@ class Admin extends Backend_Controller {
 
 		if ($this->form_validation->run('posts') == TRUE)
 		{
+			$file_data = NULL;
+
 			if (!empty($_FILES['userfile']['name']))
 			{
 				$config['upload_path']   = $this->folder_path;
@@ -119,21 +121,15 @@ class Admin extends Backend_Controller {
 				$config['file_name']     = $id;
 				$this->upload->initialize($config);
 
-				if ($this->upload->do_upload('userfile'))
-				{
-					$file_data = $this->upload->data();
-				}
-				else {
-					$this->session->set_flashdata('create_error', $this->upload->display_errors('', ''));
-                	$file_data = NULL;
-				}
+				if ($this->upload->do_upload('userfile')) $file_data = $this->upload->data();
+				else $this->session->set_flashdata('create_error', $this->upload->display_errors('', ''));
 			}
 
 			// Found new file delete the old one
 			$fileBanner = $this->posts_m->get($id)->image;
 			if(!empty($file_data))
 			{
-				unlink($this->folder_path.$fileBanner);
+				if(file_exists($this->folder_path.$fileBanner)) unlink($this->folder_path.$fileBanner);
 				$fileBanner = $file_data['file_name'];
 			}
